backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
motor==3.6.0
pymongo==4.9.0
pydantic==2.9.2
pydantic-settings==2.5.2
pyhdfs==0.3.1
loguru==0.7.2
redis==5.1.0
tableau-api-lib==0.1.45
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==9.0.0
pandas==2.2.3
scikit-learn==1.5.2
python-dotenv==1.0.1